Fixed Circus Tent, Cube now Jumps
I added left and right keys as also up and down movement because intuitively, when I first started to play the game, I wanted to click right and left keys to control the cube, so I thought if players skip the controls it might be easy for them to just go to left and right buttons to play. I added a jumping mechanism to the cube, which has a "JumpHeight" variable in the move script if anyone thinks the jumping is oddly high, but it's also neat that you can control your jumping by how long you push the space button. Double jumps are there as well (by the magic of unity?).